requirements.txt
django>=3.0.6,<3.1.1
djangorestframework>=3.11.0,<3.12.0
flake8>=3.8.1,<3.9.0
psycopg2>=2.8.5<2.9.0